refactor: update latitude and longitude types from int to double in AddRoomRequest model

+++ lib/Features/add_property/data/models/add_room_request.dart
@@ -69,8 +69,8 @@ class AddRoomRequest {
   String? street;
   String? city;
   String? government;
-  int? latitude;
-  int? longitude;
+  double? latitude;
+  double? longitude;
   List<SharedPropertyImages>? propertyImages;
   List<Rooms>? rooms;
   bool? isDraft;
